// src/app/models/device.ts
export class Device {

  id: number;
  avatar?: string;
  dvdvid: string;
  dvbrch: string;
  dvdown?:string;
  dvdrch?:string;
  dvloc1?: string;
  dvloc2?: string;
  dvloc3?: string;
  dvserl?: string;
  dvmake?: string;
  dvmodl?: string;
  latitude?: string;
  longitude?: string;
  report?: string;
  userid?: string;
  createdAt?: string;
  updateAt?: string;
}